Add Classes & IDs
In this project I will be using classes and IDs in CSS. All odd number
elements will have a light pink background, with a list of fonts. The
second element will have blue text and a font size of 36px. The third
element will have what every odd number element styles in addition a
font size of 24px. The fourth element will have a light green
background with a font size of 24px and the text will be bold.

This lesson has taught me in the end to not use repetitive code.
Noticing that there will be 2 instances where I have to change the font
size to 24px, this gives me the opportunity to add a class styling to
an element that already has an ID for styling.