Fix Cascade
In this lesson I had to edit and switch some selectors around to fix
specificity issues. I made my project to be as identical to the desired
outcome photo. I noticed that there were slight differences in the photo
compared to the solution code. The word 'text' goes out further than the
cancel button. In the solution you will find that same text to be within
both buttons.

Other than that small difference, I feel that this project was a bit
challenging and I may need to work on my understanding more when it
comes to specificity.