Add Margin & Padding
This exercise was easy for me to understand and completed. The only
part I struggled with is the margin-left property. I placed my values
at 335px to right align DIV THREE, but when looking at the solution
when I completed the lesson made me understand more why auto was used
instead.

So far I am understanding what differences I make compared to the
solutions provided in this course. And noticing my progression has
given me a great sense of accomplishment.